Changed behaviour of 'isFilePath'
When Imagick is installed and used as the driver it can work with a
pageselect in the imagepath, note that this only works if ghostscript
is installed. But it then allows the library to quickly load a single
page instead of having to load the entire document; therefor
increasing the parsing speed of the document quite dramatically. When
ghostscript is not installed it will not break when imagick is used.

We also added another testcase to make sure this works. To get this
to work we had to mock the 'Config' facade, which has been done using
a mock proxy.