Align the behaviour of Imagick crop command with GD
The GD crop command creates a new canvas for when cropping. The upshot is
that if you pass the start position -x, the image starts at x on the new
canvas. ImageMagick does not have this behaviour in crop, which starts
strictly at 0,0. extentImage however will duplicate the behaviour of the
crop command in the GD driver, meaning both now perform exactly the same
with negative input values.