Change encoding logic
The logic for encoding image data in connection with a limit on total
colors has been removed. The number of colors of an image can now be
reduced with the corresponding modifier, but this is not related to the
output of the encoder.

Furthermore, there is now an AutoEncoder, which resumes the original
format of the original image source and selects the encoder accordingly.
This Origin attribute is set during the decoding process so that it can
be read out again later. The AutoEncoder is now the default value of the
parameter for Image::encode().