Check if outline stroke effect colors are applicable
The text outline effect is drawn with a trick by plotting additional
text under the actual text with an offset in the color of the outline
effect. For this reason, no colors with transparency can be used for the
text color or the color of the stroke effect, as this would be
superimposed.