Add text stroke outline effect
Adds the possibility to draw text with an outline stroke effect. The color and width of the outline can be specified.

During development, it was noticed that Imagick can handle the effect natively, but always draws it in the middle of the border. As there is currently no option to change this, the same trick was used that was already used for the implementation with the GD library.

This "compound" method first draws the outline several times with an offset before the actual text is placed over it.

However, this has the disadvantage that no transparent colors can be used if the stroke/outline effect is active as the would superimpose each other.

---------

Co-authored-by: MaximusLight <maximuslight7@gmail.com>
Co-authored-by: Oliver Vogel <oliver@olivervogel.com>
Co-authored-by: Amowogbaje Gideon <amowogbajegideon@gmail.com>